Default to using programs@sparksf for contact owner. Dont update contact owner if set
There have been intermittent failures with contact updates with the error message 'owner is inactive, cannot reparent record," though this seems to occur even when the owner is set to an active user. Refraining from trying to update this field should help, and perhaps setting the owner to be a more "neutral" user will help.